+++ src/utils/applyAnimationToNode/index.ts
@@ -1,1 +1,1 @@
-export * from './applyAnimationToNode';
+export { applyAnimationToNode } from './applyAnimationToNode';
