+++ src/utils/generateAnimation/index.ts
@@ -1,1 +1,1 @@
-export * from './generateAnimation';
+export { generateAnimation } from './generateAnimation';
