--- src/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './applyAnimationToNode';
-export * from './generateAnimation';
-export * from './splitText';
-export * from './validation';
