+++ src/utils/validation/index.ts
@@ -1,2 +1,2 @@
-export * from './handleValidation';
-export * from './validation';
+export { handleValidation } from './handleValidation';
+export { validateNodeMotionProps, validateTextMotionProps } from './validation';
